refactor rules into separate classes
