+++ infra/growth-event/src/enums.js
@@ -33,7 +33,9 @@ const GrowthEventTypes = new Enum(
   'ListingSold', // Seller side event.
   'MobileAccountCreated', // User installed Origin mobile app and registered an account.
   'SharedOnTwitter',
-  'FollowedOnTwitter'
+  'FollowedOnTwitter',
+  'SharedOnFacebook',
+  'FollowedOnFacebook'
 )
 
 module.exports = {
