+++ infra/growth/src/resources/rules/multiEventsRule.js
@@ -1,0 +1,70 @@
+const { BaseRule } = require('./baseRule')
+
+/**
+ * A rule that requires N events out of a list of event types.
+ *
+ */
+class MultiEventsRule extends BaseRule {
+  constructor(crules, levelId, config) {
+    super(crules, levelId, config)
+
+    if (!this.config.eventTypes) {
+      throw new Error(`${this.str()}: missing eventTypes field`)
+    }
+    this.config.eventTypes.forEach(eventType => this.addEventType(eventType))
+
+    if (
+      !this.config.numEventsRequired ||
+      !Number.isInteger(this.config.numEventsRequired) ||
+      this.config.numEventsRequired > this.eventTypes.length
+    ) {
+      throw new Error(`${this.str()}: missing or invalid numEventsRequired`)
+    }
+    this.numEventsRequired = this.config.numEventsRequired
+  }
+
+  /**
+   * Returns number of rewards user qualifies for, taking into account the rule's limit.
+   * @param {string} ethAddress - User's account.
+   * @param {Array<models.GrowthEvent>} events
+   * @returns {number}
+   * @private
+   */
+  async _numRewards(ethAddress, events) {
+    // Attempts to picks N different events from the tally.
+    // Returns true if success, false otherwise.
+    function pickN(tally, n) {
+      let numPicked = 0
+      for (const key of Object.keys(tally)) {
+        if (tally[key] > 0) {
+          tally[key]--
+          numPicked++
+        }
+        if (numPicked === n) {
+          break
+        }
+      }
+      return numPicked === n
+    }
+
+    const tally = this._tallyEvents(ethAddress, this.eventTypes, events)
+    let numRewards = 0
+    while (numRewards < this.limit && pickN(tally, this.numEventsRequired)) {
+      numRewards++
+    }
+    return numRewards
+  }
+
+  /**
+   * Calculates if the rule passes.
+   * @param {string} ethAddress - User's account.
+   * @param {Array<models.GrowthEvent>} events
+   * @returns {boolean}
+   */
+  async _evaluate(ethAddress, events) {
+    const tally = this._tallyEvents(ethAddress, this.eventTypes, events)
+    return Object.keys(tally).length >= this.numEventsRequired
+  }
+}
+
+module.exports = { MultiEventsRule }
