+++ infra/growth/src/resources/rules/reward.js
@@ -1,0 +1,10 @@
+class Reward {
+  constructor(campaignId, levelId, ruleId, value) {
+    this.campaignId = campaignId
+    this.levelId = levelId
+    this.ruleId = ruleId
+    this.value = value // <{amount: string, currency: string}>
+  }
+}
+
+module.exports = { Reward }
